tests/wpt/web-platform-tests/trusted-types/block-string-assignment-to-Document-write.tentative.html

<!DOCTYPE html>
<html>
<head>
  <script src="/resources/testharness.js"></script>
  <script src="/resources/testharnessreport.js"></script>
  <script src="support/helper.sub.js"></script>

  <meta http-equiv="Content-Security-Policy" content="trusted-types *">
</head>
<body>
<script>
  // String assignments throw.
  test(t => {
    assert_throws(new TypeError(), _ => {
      document.write('A string');
    });
  }, "`document.write(string)` throws");

  // Null assignment throws.
  test(t => {
    assert_throws(new TypeError(), _ => {
      document.write(null);
    });
  }, "`document.write(null)` throws");
</script>
